COBRA-3020 - added CORS filters, added update to attach filters

+++ plugins/filters/index.js
@@ -11,7 +11,11 @@ ${Object.keys(filter.options).map((opt) => {
   let label = opt.replace('_', ' ').split(' ').map((x) => {
     return x[0].toUpperCase() + x.substring(1)
   }).join(' ')
-  return '  ***' + label + ':*** ' + filter.options[opt];
+  if(Array.isArray(filter.options[opt])) {
+    return '  ***' + label + ':*** ' + filter.options[opt].join(", ");
+  } else {
+    return '  ***' + label + ':*** ' + filter.options[opt];
+  }
 }).join('\n')}
   `;
 }
@@ -20,6 +24,16 @@ function format_filter_attachment(fa) {
   return `**Ⴤ ${fa.id}**
   ***Type:*** ${fa.filter.type}
   ***Filter:*** ${fa.filter.name}
+${Object.keys(fa.filter.options).map((opt) => { 
+  let label = opt.replace('_', ' ').split(' ').map((x) => {
+    return x[0].toUpperCase() + x.substring(1)
+  }).join(' ')
+  if(Array.isArray(fa.filter.options[opt])) {
+    return '  ***' + label + ':*** ' + fa.filter.options[opt].join(", ");
+  } else {
+    return '  ***' + label + ':*** ' + fa.filter.options[opt];
+  }
+}).join('\n')}
 ${Object.keys(fa.options).map((opt) => { 
   let label = opt.replace('_', ' ').split(' ').map((x) => {
     return x[0].toUpperCase() + x.substring(1)
@@ -51,12 +65,61 @@ async function create(appkit, args) {
         task.end('error');
         return appkit.terminal.error(new Error('A JWT http filter requires options "jwt-jwks-uri" and "jwt-issuer" be set. It may optionally have one (or more) "audiences".'))
       }
-      options = {jwks_uri:args['jwt-jwks-uri'], issuer:args['jwt-issuer'], audiences:args['jwt-audiences']}
+      options = {
+        "jwks_uri":args['jwt-jwks-uri'], 
+        "issuer":args['jwt-issuer'], 
+        "audiences":args['jwt-audiences'],
+      }
+    } else if (args.type === "cors") {
+      options = {
+        "allow_origin":args['cors-allow-origin'],
+        "allow_methods":args['cors-allow-methods'],
+        "allow_headers":args['cors-allow-headers'],
+        "expose_headers":args['cors-expose-headers'],
+        "max_age":args['cors-max-age'],
+        "allow_credentials":args['cors-allow-credentials'],
+      }
     } else {
       task.end('error');
       return appkit.terminal.error(new Error('The specified filter type was invalid, the supported options are: jwt'))
     }
     await appkit.api.post(JSON.stringify({"type":args.type, "name":args.FILTER_NAME, options, description:args.description, organization:args.org}),`/filters`)
+  } catch (e) {
+    task.end('error');
+    return appkit.terminal.error(e);
+  }
+  task.end('ok')
+}
+
+async function update(appkit, args) {
+  let task = appkit.terminal.task(`Updating http filter **⬢ ${args.FILTER_NAME}** ###(changes take place when an attached app deploys)###`);
+  task.start()
+  try {
+    let options = {}
+    if(args.type === "jwt") {
+      if(!args['jwt-jwks-uri'] || !args['jwt-issuer']) {
+        task.end('error');
+        return appkit.terminal.error(new Error('A JWT http filter requires options "jwt-jwks-uri" and "jwt-issuer" be set. It may optionally have one (or more) "audiences".'))
+      }
+      options = {
+        "jwks_uri":args['jwt-jwks-uri'], 
+        "issuer":args['jwt-issuer'], 
+        "audiences":args['jwt-audiences'],
+      }
+    } else if (args.type === "cors") {
+      options = {
+        "allow_origin":args['cors-allow-origin'],
+        "allow_methods":args['cors-allow-methods'],
+        "allow_headers":args['cors-allow-headers'],
+        "expose_headers":args['cors-expose-headers'],
+        "max_age":args['cors-max-age'],
+        "allow_credentials":args['cors-allow-credentials'],
+      }
+    } else {
+      task.end('error');
+      return appkit.terminal.error(new Error('The specified filter type was invalid, the supported options are: jwt'))
+    }
+    await appkit.api.put(JSON.stringify({"type":args.type, "name":args.FILTER_NAME, options, description:args.description, organization:args.org}),`/filters/${args.FILTER_NAME}`)
   } catch (e) {
     task.end('error');
     return appkit.terminal.error(e);
@@ -87,10 +150,6 @@ async function destroy(appkit, args) {
   }
 }
 
-function update(appkit, args) {
-  // TODO
-}
-
 async function attach(appkit, args) {
   let attach = async (input) => {
     if(input !== args.app) {
@@ -103,6 +162,9 @@ async function attach(appkit, args) {
       let options = {}
       if (args.excludes && Array.isArray(args.excludes)) {
         options.excludes = args.excludes
+      }
+      if (args.includes && Array.isArray(args.includes)) {
+        options.includes = args.includes
       }
       await appkit.api.post(JSON.stringify({"filter":{"id":filter_info.id}, options}),`/apps/${args.app}/filters`);
     } catch (e) {
@@ -118,17 +180,58 @@ async function attach(appkit, args) {
   }
 }
 
-async function detach(appkit, args) {
-  let task = appkit.terminal.task(`Detaching **Ⴤ ${args.FILTER_ATTACHMENT_ID}**`);
-  task.start();
-  try {
-    await appkit.api.delete(`/apps/${args.app}/filters/${args.FILTER_ATTACHMENT_ID}`);
-  } catch (e) {
-    task.end('error');
-    return appkit.terminal.error(e);
+async function attach_update(appkit, args) {
+  let attach_update = async (input) => {
+    if(input !== args.app) {
+      return appkit.terminal.soft_error(`Confirmation did not match !!${args.app}!!. Aborted.`);
+    }
+    let task = appkit.terminal.task(`Updating http filter attachment **Ⴤ ${args.FILTER_ATTACHMENT_ID}** on ##⬢ ${args.app}##`);
+    task.start()
+    try {
+      let filter_info = await appkit.api.get(`/apps/${args.app}/filters/${args.FILTER_ATTACHMENT_ID}`)
+      let options = {}
+      if (args.excludes && Array.isArray(args.excludes)) {
+        options.excludes = args.excludes
+      }
+      if (args.includes && Array.isArray(args.includes)) {
+        options.includes = args.includes
+      }
+      await appkit.api.put(JSON.stringify({"filter":{"id":filter_info.filter.id}, options}),`/apps/${args.app}/filters/${args.FILTER_ATTACHMENT_ID}`);
+    } catch (e) {
+      task.end('error');
+      return appkit.terminal.error(e);
+    }
+    task.end('ok')
   }
-  task.end('ok')
+  if(args.confirm) {
+    await attach(args.confirm);
+  } else {
+    appkit.terminal.confirm(` ~~▸~~    !!DANGER ZONE!!: This feature is still in beta, and updating this filter attachment on **⬢ ${args.app}** may result in instability.\n ~~▸~~    Before continuing ensure you've read ##https://github.com/akkeris/akkeris/issues/9## and have implemented its recommendations.\n ~~▸~~    To proceed, type !!${args.app}!! or re-run this command with !!--confirm ${args.app}!!\n`, attach_update);
+  }
 }
+
+async function detach(appkit, args) {
+  let del = async (input) => {
+    if(input !== args.app) {
+      return appkit.terminal.soft_error(`Confirmation did not match !!${args.app}!!. Aborted.`);
+    }
+    let task = appkit.terminal.task(`Detaching **Ⴤ ${args.FILTER_ATTACHMENT_ID}**`);
+    task.start();
+    try {
+      await appkit.api.delete(`/apps/${args.app}/filters/${args.FILTER_ATTACHMENT_ID}`);
+    } catch (e) {
+      task.end('error');
+      return appkit.terminal.error(e);
+    }
+    task.end('ok')
+  };
+  if(args.confirm) {
+    await del(args.confirm);
+  } else {
+    appkit.terminal.confirm(` ~~▸~~    WARNING: This will detach **⬢ ${args.FILTER_ATTACHMENT_ID}** from app ${args.app}.\n ~~▸~~    To proceed, type !!${args.app}!! or re-run this command with !!--confirm ${args.app}!!\n`, del);
+  }
+}
+
 module.exports = {
   init:function(appkit) {
     let confirm_option = {
@@ -138,6 +241,7 @@ module.exports = {
         'description':'The confirmation string to use when removing'
       }
     }
+
     let filters_create_option = {
       'description':{
         'alias':'d',
@@ -153,7 +257,7 @@ module.exports = {
       },
       'type':{
         'alias':'t',
-        'choices':['jwt'],
+        'choices':['jwt','cors'],
         'demand':true,
         'description':'The type of http filter.'
       },
@@ -167,14 +271,101 @@ module.exports = {
       },
       'jwt-audiences':{
         'array':true,
-        'description':'One or more jwt audiences to sue for a JWT oauth filter.'
+        'description':'One or more jwt audiences to use for a JWT oauth filter (optional if type=jwt).'
+      },
+      'cors-allow-origin':{
+        'array':true,
+        'description':'The allowed origins for the CORS filter (optional if type=cors)'
+      },
+      'cors-allow-methods':{
+        'array':true,
+        'description':'The allowed HTTP methods for the CORS filter (optional if type=cors)'
+      },
+      'cors-expose-headers':{
+        'array':true,
+        'description':'The allowed HTTP headers for the CORS filter (optional if type=cors)'
+      },
+      'cors-allow-headers':{
+        'array':true,
+        'description':'The allowed HTTP headers in the request (optional if type=cors)'
+      },
+      'cors-max-age':{
+        'number':true,
+        'description':'The age (in seconds) the CORS filter may be cached (optional if type=cors)'
+      },
+      'cors-allow-credentials':{
+        'boolean':true,
+        'description':'Whether credentials are allowed to be sent in the CORS filter (optional if type=cors)',
       }
     }
+
+    let filters_update_option = {
+      'description':{
+        'alias':'d',
+        'string':true,
+        'demand':true,
+        'description':'The description used for the http filter.'
+      },
+      'org':{
+        'alias':'o',
+        'string':true,
+        'demand':true,
+        'description':'The name of the organization who owns this filter.'
+      },
+      'type':{
+        'alias':'t',
+        'choices':['jwt','cors'],
+        'demand':true,
+        'description':'The type of http filter.'
+      },
+      'jwt-issuer':{
+        'string':true,
+        'description':'The issuer to use for a JWT oauth filter (required if type=jwt).'
+      },
+      'jwt-jwks-uri':{
+        'string':true,
+        'description':'The jwks uri to use for a JWT oauth filter (required if type=jwt).'
+      },
+      'jwt-audiences':{
+        'array':true,
+        'description':'One or more jwt audiences to use for a JWT oauth filter (optional if type=jwt).'
+      },
+      'cors-allow-origin':{
+        'array':true,
+        'description':'The allowed origins for the CORS filter (optional if type=cors)'
+      },
+      'cors-allow-methods':{
+        'array':true,
+        'description':'The allowed HTTP methods for the CORS filter (optional if type=cors)'
+      },
+      'cors-expose-headers':{
+        'array':true,
+        'description':'The exposed HTTP headers returned in the response (optional if type=cors)'
+      },
+      'cors-allow-headers':{
+        'array':true,
+        'description':'The allowed HTTP headers in the request (optional if type=cors)'
+      },
+      'cors-max-age':{
+        'number':true,
+        'description':'The age (in seconds) the CORS filter may be cached (optional if type=cors)'
+      },
+      'cors-allow-credentials':{
+        'boolean':true,
+        'description':'Whether credentials are allowed to be sent in the CORS filter (optional if type=cors)',
+      }
+    }
+
     let filters_attach = {
       'excludes':{
         'alias':'e',
         'array':true,
-        'description':'One or more relative path prefixes to exclude from the filter (e.g., /test would exclude /test/foo and /test).'
+        'description':'One or more relative path prefixes to exclude from the filter (availble only on JWT filters).'
+      },
+      'includes':{
+        'alias':'i',
+        'array':true,
+        'description':'One or more relative path prefixes to explicitly include (otherwise everything is included, only on JWT filters).'
       },
       'app':{
         'alias':'a',
@@ -189,6 +380,7 @@ module.exports = {
         'description':'Confirm with the application name that you want to continue attaching this filter.'
       }
     }
+
     let require_app_option = {
       'app':{
         'alias':'a',
@@ -197,6 +389,7 @@ module.exports = {
         'description':'The app to act on'
       }
     }
+
     let optional_app = {
       'app':{
         'alias':'a',
@@ -209,10 +402,12 @@ module.exports = {
     appkit.args
       .command('filters', 'List available http filters that can be attached.', optional_app, list.bind(null, appkit))
       .command('filters:create FILTER_NAME', 'Create a new http filter', filters_create_option, create.bind(null, appkit))
+      .command('filters:update FILTER_NAME', 'Update an existing http filter', filters_update_option, update.bind(null, appkit))
       .command('filters:destroy FILTER_NAME', 'Destroy an http filter', confirm_option, destroy.bind(null, appkit))
       //.command('filters:update FILTER_NAME [options..]', 'Update an http filter', {}, update.bind(null, appkit))
       .command('apps:filters:attach FILTER_NAME', 'Attach an http filter to an app', filters_attach, attach.bind(null, appkit))
-      .command('apps:filters:detach FILTER_ATTACHMENT_ID', 'Attach an http filter to an app', require_app_option, detach.bind(null, appkit))
+      .command('apps:filters:detach FILTER_ATTACHMENT_ID', 'Detach an http filter to an app', {...require_app_option, ...confirm_option}, detach.bind(null, appkit))
+      .command('apps:filters:update FILTER_ATTACHMENT_ID', 'Update an http filter attachment on an app', filters_attach, attach_update.bind(null, appkit))
       .command('apps:filters', 'List http filters currently on an app', require_app_option, list_attachments.bind(null, appkit))
       // aliases
       .command('filters:attach FILTER_NAME', false, filters_attach, attach.bind(null, appkit))
